refactor: separate json schema for fast-json-stringify
