+++ src/output_utils.ts
@@ -1,31 +1,5 @@
-const fastJson = require('fast-json-stringify')
-
-type ReplacerFunction = (key: string | number, value: unknown) => unknown
-
-const logSchema = {
-    title: 'Log',
-    type: 'object',
-    properties: {
-        level: { type: 'string' },
-        time: {
-            type: ['string', 'number', 'null'],
-        },
-        namespace: { type: 'string' },
-        contextId: { type: ['string', 'null'] },
-        message: { type: ['string', 'null'] },
-        meta: {
-            type: ['object', 'null'],
-            additionalProperties: true,
-        },
-        data: {
-            type: ['object', 'null'],
-            additionalProperties: true,
-        },
-    },
-    additionalProperties: true,
-}
-
-const fastStringifyLog = fastJson(logSchema)
+import type { ReplacerFunction } from './definitions'
+import { fastStringifyLog } from './json_schema'
 
 /**
  * Replace circular reference when used with JSON.stringify
